Add file upload and move more code to Vuex
User can now upload a CSV file and have the plot update in real time
(therefore the "Get Plot" button is no longer necessary).

Store more values in the Vuex store and move the logic that computes
them over there as well, as it's easier to handle reactivity and
synchronization that way.

New dependency: PapaParse, for parsing CSV files.