Add disabling buttons and marking them as active
Buttons that should not be clicked because the information they would
display is not available now appear disabled, and become enabled once
the relevant data is available.

In the "change shown values" and "change plot" menus, the "active"
button, i.e. the button whos values are being displayed on screen,
appears "glossy" - a different color.