Rewrite DataInputForm
Make it so that only one file can be uploaded at a time.

This is a better framework if dealing with operations that can
take very long to compute, as it prevents all steps
(calibration, etc.) from being executed after the submit button
is pressed, which can take very long.

It's also easier to deal with keeping values up to date.