Change ui store valuesStatus from strings to bools
It's only responsability is to inform whether certain values are
available or not. This is because only one request will be processed at
a time, so there is no possibility of having two sets of values loading
at the same time.

There will be just one state variable to inform if values are being
loaded or not, and which are being loaded is irrelevant.